+++ index.d.ts
@@ -36,6 +36,7 @@ export type FormApi = {
   getFieldState: (name: string) => FieldState;
   validate: () => void;
   validateField: (name: string) => void;
+  submitForm: () => void;
 };
 
 export type FieldState = {
@@ -75,8 +76,25 @@ export type FieldApi = {
   getMaskedValue: () => unknown;
 };
 
+export type MultistepApi = {
+  next: () => void;
+  previous: () => void;
+  getNextStep: () => void;
+  getPreviousStep: () => string;
+  setCurrent: (name: string) => string;
+  getCurrentStep: () => string;
+};
+
+export type MultistepState = {
+  steps: Array<string>;
+  goal: string;
+  current: string;
+  nextStep: string;
+  previousStep: string;
+};
+
 export type ArrayFieldApi = {
-  add: () => void;
+  add: (amount?: number) => void;
   reset: () => void;
   swap: (a: number, b: number) => void;
   addWithInitialValue: (value: any) => void;
@@ -249,6 +267,8 @@ export function useScope(name: string): string;
 
 export function useScoper(name: string): (name: string) => string;
 
+export function useMultistepApi(): MultistepApi;
+
 export function useForm<UserProps>(
   formProps: InformedProps<UserProps>
 ): {
@@ -302,6 +322,16 @@ declare namespace ArrayField {
   }: {
     children: (props: ArrayFieldItemApi & ArrayFieldItemInfo) => JSX.Element;
   }): JSX.Element;
+}
+
+declare function Multistep({
+  children
+}: {
+  children: (props: MultistepApi & MultistepState) => JSX.Element;
+}): JSX.Element;
+
+declare namespace Multistep {
+  function Step({ children }: { children: () => JSX.Element }): JSX.Element;
 }
 
 declare namespace utils {
@@ -392,7 +422,8 @@ declare namespace utils {
   function informedParse(val: any, parser: any): any;
   function informedFormat(
     val: any,
-    frmtr: any
+    frmtr: any,
+    old?: any
   ):
     | {
         value: any;
@@ -404,7 +435,8 @@ declare namespace utils {
       };
   function informedFormatter(
     val: any,
-    frmtr: any
+    frmtr: any,
+    old?: any
   ): {
     value: any;
     offset: number;