+++ src/Context.js
@@ -14,3 +14,4 @@ export const ArrayFieldItemStateContext = React.createContext();
 export const RelevanceContext = React.createContext();
 export const RadioGroupContext = React.createContext();
 export const ScopeContext = React.createContext();
+export const InformedContext = React.createContext();