+++ src/hooks/useCursorPosition.js
@@ -46,31 +46,31 @@ const useCursorPosition = ({
     () => {
       if (!inputRefs && inputRef.current != null && getCursor()) {
         // inputRef.current.selectionEnd = getCursor() + getCursorOffset();
-        const cursorLoc = getCursor() + getCursorOffset();
+        const offset = getCursorOffset() < 0 ? 0 : getCursorOffset();
+        const cursorLoc = getCursor() + offset;
+        // console.log('Cursor', getCursor());
+        // console.log('Offset', offset);
+        // console.log('CursorLoc', cursorLoc);
         // Only offset if we need to
-        if (getCursorOffset() != 0 && maintainCursor) {
-          inputRef.current.setSelectionRange(cursorLoc + 1, cursorLoc + 1);
-        } else if (maintainCursor) {
+        if (maintainCursor) {
           inputRef.current.setSelectionRange(cursorLoc, cursorLoc);
         }
       }
       if (inputRefs && getCursor()) {
+        // console.log('Cursor', getCursor());
         // Loop over each ref and set selection range
         Object.keys(inputRefs).forEach(key => {
-          const cursorLoc = getCursor()[key] + getCursorOffset()[key];
+          const offset =
+            getCursorOffset()[key] < 0 ? 0 : getCursorOffset()[key];
+          const cursorLoc = getCursor()[key] + offset;
           // Only offset if we need to
-          if (getCursorOffset()[key] != 0 && maintainCursor) {
-            inputRefs[key].current.setSelectionRange(
-              cursorLoc + 1,
-              cursorLoc + 1
-            );
-          } else if (maintainCursor) {
+          if (maintainCursor) {
             inputRefs[key].current.setSelectionRange(cursorLoc, cursorLoc);
           }
         });
       }
     },
-    [value]
+    [value, cursor]
   );
 
   return {