+++ src/hooks/useScopedApi.js
@@ -19,7 +19,8 @@ const buildScopedFormApi = (scope, formApi) => {
     validate: field => formApi.validate(`${scope}.${field}`),
     getDirty: field => formApi.getDirty(`${scope}.${field}`),
     getPristine: field => formApi.getPristine(`${scope}.${field}`),
-    getMaskedValue: field => formApi.getMaskedValue(`${scope}.${field}`)
+    getMaskedValue: field => formApi.getMaskedValue(`${scope}.${field}`),
+    clearValue: field => formApi.clearValue(`${scope}.${field}`)
   };
 };