+++ stories/Arrays/DependentFields/index.js
@@ -70,13 +70,13 @@ const CarOrTruck = () => {
 
 const ProductSelect = () => {
   const { value, dirty } = useFieldState('type');
-  const { reset } = useFieldApi('product');
+  const { clearValue } = useFieldApi('product');
 
   const opts = useMemo(() => options[value] || [], [value]);
 
   useEffect(
     () => {
-      if (dirty) reset();
+      if (dirty) clearValue();
     },
     [value]
   );