+++ stories/Formatting/NumberFormatter/index.js
@@ -1089,14 +1089,18 @@ const FormattedField = () => {
     [currency, locale]
   );
 
+  // const { formatter, parser } = useMemo(() => {
+  //   return utils.createIntlNumberFormatter('en-US', {});
+  // }, []);
+
   return (
     <Input
       field="localeMask"
-      label="Locale Masked Field (EUR currency)"
+      label={`Locale Masked Field (${currency} currency)`}
       formatter={formatter}
       parser={parser}
       formatterDependencies={[locale, currency]}
-      initialValue={3000}
+      initialValue={3000.25}
     />
   );
 };
@@ -1108,13 +1112,13 @@ const FormatParse = () => (
         label="Locale"
         field="locale"
         options={localeOptions}
-        initialValue="nl-NL"
+        initialValue="en-US"
       />
       <Select
         label="Currency"
         field="currency"
         options={currencyOptions}
-        initialValue="EUR"
+        initialValue="USD"
       />
       <FormattedField />
       <button type="submit">Submit</button>